Split off the api module into an annotations module
Both the api and the compiler module now depend on the annotations
module. This removes the need for the android-23.jar dependency, as
well as the dual sourceSets configuration in compiler’s build.gradle.
