+++ annotations/build.gradle
@@ -1,0 +1,3 @@
+apply plugin: 'java'
+
+sourceCompatibility = 1.7
