+++ annotations/src/main/java/io/sweers/barber/Config.java
@@ -1,0 +1,6 @@
+package io.sweers.barber;
+
+public class Config {
+
+    public static final String SUFFIX = "$$Barbershop";
+}
