+++ annotations/src/main/java/io/sweers/barber/Required.java
@@ -1,0 +1,15 @@
+package io.sweers.barber;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Annotate a field with @Required to indicate to Barber (and future users) that this attribute must be specified
+ */
+@Target({ElementType.FIELD, ElementType.METHOD})
+@Retention(RetentionPolicy.SOURCE)
+public @interface Required {
+}
