+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':sample', ':compiler', ':api'
+include ':sample', ':compiler', ':api', ':annotations'
